# static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Datilia API",
    "version": "1.0.0",
    "description": "## What is Datilia?\n\nDatilia is a prototype tool developed by [Chequeado](https://chequeado.com) for the **Media Party × World Bank Data360 Global Challenge**. It takes journalistic text, automatically extracts verifiable statistical claims, cross-references them against the [World Bank Data360](https://data360.worldbank.org) database, and returns structured data context with a narrative and a Vega-Lite chart.\n\n**Live demo:** [datilia.chequeado.com](https://datilia.chequeado.com) — user: `data360` / password: `data360challenge`\n",
    "contact": {
      "name": "Chequeado",
      "url": "https://chequeado.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/chequeado/datilia/blob/master/LICENSE"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8090",
      "description": "Local Docker (default)"
    }
  ],
...
  "paths": {
...
    "/contextualize": {
      "post": {
        "tags": ["Contextualization"],
        "summary": "Contextualize a single claim",
        "description": "Runs the full Datilia pipeline for a single claim:\n1. An LLM agent searches Data360 via MCP tools (up to 8 tool turns)\n2. Extracted data context is structured and persisted\n3. An LLM selects chart type and parameters\n4. A Vega-Lite chart spec is built\n5. A narrative paragraph is generated\n\nThe result is persisted as a `ContextualizationRun` and is addressable by `trace_id`.",
        "operationId": "contextualize",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ContextualizeRequest" },
              "example": {
                "claim": "Argentina's poverty rate reached 40.1% in 2023",
                "context": "Paragraph from an article about social policy in Latin America.",
                "language": "es"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contextualization completed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ContextualizeResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContextualizeRequest": {
        "type": "object",
        "required": ["claim"],
        "properties": {
          "claim": {
            "type": "string",
            "description": "The statistical or factual claim to contextualize",
            "example": "Global CO2 emissions reached a record 37.4 billion tonnes in 2023"
          },
          "context": {
            "type": "string",
            "nullable": true,
            "description": "Optional surrounding text providing context for the claim (improves indicator matching)"
          },
          "language": {
            "type": "string",
            "enum": ["es", "en", "pt"],
            "default": "es",
            "description": "Language for the narrative output"
          }
        }
      },
...
      "ChartParams": {
        "type": "object",
        "properties": {
          "strategy": {
            "type": "string",
            "enum": ["top_k", "top_k_others", "line", "line_dots", "cross_sectional", "distribution", "breakdown_comparison", "small_multiples", "stacked_bar"],
            "example": "line",
            "description": "Chart strategy selected by the AI chart selector."
          },
          "x_field": { "type": "string", "example": "time_period" },
          "y_field": { "type": "string", "example": "value" },
          "color_field": { "type": "string", "nullable": true, "example": "ref_area" },
          "facet_field": { "type": "string", "nullable": true },
          "highlight": {
            "type": "array",
            "items": { "type": "string" },
            "nullable": true,
            "example": ["ARG", "BRA"],
            "description": "ISO-3 country codes to visually emphasize; everything else is grayed out."
          },
          "top_n": { "type": "integer", "nullable": true, "example": 10 }
        }
      },
      "ToolCall": {
        "type": "object",
        "properties": {
          "tool": { "type": "string", "example": "data360_search_indicators" },
          "arguments": { "type": "object" },
          "result": {}
        }
      },
      "ContextualizeResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "enum": ["ok", "error"] },
          "trace_id": { "type": "string", "format": "uuid" },
          "is_verifiable": {
            "type": "boolean",
            "description": "Whether the claim could be matched to a Data360 indicator"
          },
          "final_text": {
            "type": "string",
            "description": "Narrative paragraph contextualizing the claim with data"
          },
          "indicator_code": { "type": "string" },
          "indicator_name": { "type": "string" },
          "database_id": { "type": "string" },
          "database_name": { "type": "string" },
          "definition": { "type": "string" },
          "periodicity": { "type": "string" },
          "unit": { "type": "string" },
          "source_url": { "type": "string", "format": "uri" },
          "area_codes": { "type": "array", "items": { "type": "string" } },
          "time_range": {
            "type": "object",
            "properties": {
              "start_year": { "type": "integer" },
              "end_year": { "type": "integer" }
            }
          },
          "chart_spec": {
            "type": "object",
            "description": "Vega-Lite chart specification (JSON)"
          },
          "chart_params": { "$ref": "#/components/schemas/ChartParams" },
          "tool_trace": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ToolCall" },
            "description": "Ordered list of MCP tool calls made by the agent"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": { "type": "string" }
        }
      }
    },
    "responses": {
      "ValidationError": {
        "description": "Request validation failed",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
      "InternalError": {
        "description": "Pipeline or server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": { "type": "string", "enum": ["error"] },
                "trace_id": { "type": "string", "format": "uuid" },
                "detail": { "type": "string" }
              }
            }
          }
        }
      },
...
    }
  }
}